refactor: rename new property class
