--- detectable/src/main/java/com/synopsys/integration/detectable/util/ListSourcedExcludedIncludedWildcardFilter.java
@@ -1,49 +1,0 @@
-/**
- * detectable
- *
- * Copyright (c) 2020 Synopsys, Inc.
- *
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.synopsys.integration.detectable.util;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import org.apache.commons.lang3.StringUtils;
-
-import com.synopsys.integration.util.ExcludedIncludedWildcardFilter;
-
-public class ListSourcedExcludedIncludedWildcardFilter extends ExcludedIncludedWildcardFilter {
-
-    public static ListSourcedExcludedIncludedWildcardFilter create(List<String> toExclude, List<String> toInclude) {
-        if (toExclude == null) {
-            toExclude = new ArrayList<>(0);
-        }
-        if (toInclude == null) {
-            toInclude = new ArrayList<>(0);
-        }
-        String toExcludeString = StringUtils.joinWith(",", toExclude.toArray());
-        String toIncludeString = StringUtils.joinWith(",", toInclude.toArray());
-        return new ListSourcedExcludedIncludedWildcardFilter(toExcludeString, toIncludeString);
-    }
-
-    private ListSourcedExcludedIncludedWildcardFilter(String toExcludeString, String toIncludeString) {
-        super(toExcludeString, toIncludeString);
-    }
-}
